+++ engine/language-server/src/main/scala/org/enso/languageserver/libraries/handler/LibraryPublishHandler.scala
@@ -19,7 +19,7 @@ import org.enso.languageserver.libraries.{
 import org.enso.languageserver.requesthandler.RequestTimeout
 import org.enso.languageserver.util.UnhandledLogging
 import org.enso.libraryupload.{auth, LibraryUploader}
-import org.enso.loggingservice.LoggingServiceManager
+import org.enso.logging.LoggingServiceManager
 
 import scala.concurrent.Future
 import scala.concurrent.duration.FiniteDuration