+++ engine/launcher/src/main/scala/org/enso/launcher/cli/LauncherApplication.scala
@@ -1,6 +1,5 @@
 package org.enso.launcher.cli
 
-import akka.http.scaladsl.model.Uri
 import cats.data.NonEmptyList
 import cats.implicits._
 import nl.gn0s1s.bump.SemVer
@@ -9,18 +8,18 @@ import org.enso.cli.arguments.Opts.implicits._
 import org.enso.cli.arguments._
 import org.enso.distribution.config.DefaultVersion
 import org.enso.distribution.config.DefaultVersion._
-import org.enso.launcher.cli.LauncherColorMode.argument
 import org.enso.launcher.distribution.DefaultManagers._
 import org.enso.launcher.installation.DistributionInstaller
 import org.enso.launcher.installation.DistributionInstaller.BundleAction
 import org.enso.launcher.upgrade.LauncherUpgrader
 import org.enso.launcher.{cli, Launcher}
-import org.enso.loggingservice.{ColorMode, LogLevel}
 import org.enso.runtimeversionmanager.cli.Arguments._
 import org.enso.runtimeversionmanager.runner.LanguageServerOptions
+import org.slf4j.event.Level
 
 import java.nio.file.Path
 import java.util.UUID
+import java.net.URI
 
 /** Defines the CLI commands and options for the program.
   *
@@ -127,12 +126,12 @@ object LauncherApplication {
   }
   private def engineLogLevel = {
     Opts
-      .optionalParameter[LogLevel](
+      .optionalParameter[Level](
         "log-level",
         "(error | warning | info | debug | trace)",
         "Sets logging verbosity for the engine. Defaults to info."
       )
-      .withDefault(LogLevel.Info)
+      .withDefault(Level.INFO)
   }
 
   private def runCommand: Command[Config => Int] =
@@ -606,14 +605,14 @@ object LauncherApplication {
       "running actions. May be needed if program output is piped.",
       showInUsage = false
     )
-    val logLevel = Opts.optionalParameter[LogLevel](
+    val logLevel = Opts.optionalParameter[Level](
       GlobalCLIOptions.LOG_LEVEL,
       "(error | warning | info | debug | trace)",
-      "Sets logging verbosity for the launcher. If not provided, defaults to" +
+      "Sets logging verbosity for the launcher. If not provided, defaults to " +
       s"${LauncherLogging.defaultLogLevel}."
     )
     val connectLogger = Opts
-      .optionalParameter[Uri](
+      .optionalParameter[URI](
         GlobalCLIOptions.CONNECT_LOGGER,
         "URI",
         "Instead of starting its own logging service, " +
@@ -627,18 +626,6 @@ object LauncherApplication {
       "variable.",
       showInUsage = false
     )
-    val colorMode =
-      Opts
-        .aliasedOptionalParameter[ColorMode](
-          GlobalCLIOptions.COLOR_MODE,
-          "colour",
-          "colors"
-        )(
-          "(auto | yes | always | no | never)",
-          "Specifies if colors should be used in the output, defaults to auto."
-        )
-        .withDefault(ColorMode.Auto)
-
     val internalOpts = InternalOpts.topLevelOptions
 
     (
@@ -650,8 +637,7 @@ object LauncherApplication {
       hideProgress,
       logLevel,
       connectLogger,
-      noLogMasking,
-      colorMode
+      noLogMasking
     ) mapN {
       (
         internalOptsCallback,
@@ -662,8 +648,7 @@ object LauncherApplication {
         hideProgress,
         logLevel,
         connectLogger,
-        disableLogMasking,
-        colorMode
+        disableLogMasking
       ) => () =>
         if (shouldEnsurePortable) {
           Launcher.ensurePortable()
@@ -673,7 +658,6 @@ object LauncherApplication {
           autoConfirm  = autoConfirm,
           hideProgress = hideProgress,
           useJSON      = useJSON,
-          colorMode    = colorMode,
           internalOptions = GlobalCLIOptions.InternalOptions(
             logLevel,
             connectLogger,
@@ -686,9 +670,7 @@ object LauncherApplication {
         LauncherLogging.setup(
           logLevel,
           connectLogger,
-          globalCLIOptions.colorMode,
-          !disableLogMasking,
-          None
+          !disableLogMasking
         )
         initializeApp()