+++ lib/scala/logging-service/src/main/java/org/enso/logging/LoggerInitializationFailed.java
@@ -1,0 +1,7 @@
+package org.enso.logging;
+
+public class LoggerInitializationFailed extends RuntimeException {
+  public LoggerInitializationFailed() {
+    super("Logger initialization failed");
+  }
+}